refactor: reorganize component structure
add server origin for static files in development
