+++ django/home/jinja2/home/index.jinja
@@ -7,7 +7,7 @@
                 <div class='col'>
                     {# Mount proof of concept vue component #}
                     <div id="app"></div>
-                    {{ vite_asset('views/home/main.ts') }}
+                    {{ vite_asset('apps/helloworld.ts') }}
                     <h1 class='text-white'>A growing collection of resources for <br>computational model-based science.</h1>
                 </div>
             </div>
