+++ frontend-next/src/apps/helloworld.ts
@@ -1,6 +1,6 @@
 import "vite/modulepreload-polyfill";
 
 import { createApp } from "vue";
-import HelloWorld from "@/views/home/HelloWorld.vue";
+import HelloWorld from "@/components/HelloWorld.vue";
 
 createApp(HelloWorld).mount("#app");
