+++ frontend-next/src/apps/main.ts
@@ -1,0 +1,3 @@
+// import global styles and bootstrap javascript
+import "@/scss/styles.scss";
+import "bootstrap";
