+++ frontend-next/src/composables/count.ts
@@ -1,0 +1,14 @@
+import { ref } from "vue"
+
+export function useCount() {
+  const count = ref(0)
+
+  const increment = () => {
+    count.value++
+  }
+
+  return {
+    count,
+    increment,
+  }
+}
