+++ frontend-next/vite.config.ts
@@ -1,19 +1,22 @@
-import { fileURLToPath, URL } from "node:url";
 import { defineConfig } from "vite";
 import vue from "@vitejs/plugin-vue";
 
 const { resolve } = require("path");
+const resolvePath = (relativePath: string) => {
+  return resolve(__dirname, relativePath);
+};
 
 export default defineConfig({
   plugins: [vue()],
-  root: resolve("./src"),
+  root: resolvePath("./src"),
   base:
     process.env.NODE_ENV === "development"
-      ? "http://localhost:3000/static/bundles/"
+      ? "http://localhost:5000/static/bundles/"
       : "/static/bundles/",
   server: {
     host: "0.0.0.0",
     port: 5000,
+    origin: "http://localhost:5000",
     watch: {
       usePolling: true,
       disableGlobbing: false,
@@ -22,7 +25,8 @@ export default defineConfig({
   resolve: {
     extensions: [".js", ".ts", ".scss", ".vue"],
     alias: {
-      "@": fileURLToPath(new URL("./src", import.meta.url)),
+      "@": resolvePath("./src"),
+      "~bootstrap": resolvePath("./node_modules/bootstrap"),
     },
   },
   build: {
@@ -30,8 +34,9 @@ export default defineConfig({
     manifest: true,
     rollupOptions: {
       input: {
-       home: resolve("./src/views/home/main.ts"),
-        // add more pages here
+        main: resolvePath("./src/main.ts"),
+        helloworld: resolvePath("./src/apps/helloworld.ts"),
+        // add more entry points here
       },
     },
   },
